プロフィール取得APIにユーザーのタイムライン情報を追加
