+++ app/domains/page/profile_page/show_user_with_timeline_domain.rb
@@ -1,0 +1,17 @@
+module Page::ProfilePage
+  class ShowUserWithTimelineDomain < ApplicationDomain
+    def initialize(user:)
+      super()
+      @user = user
+    end
+
+    def execute
+      posts = @user.posts.includes(:likes, :replies).order(created_at: :desc).limit(20)
+
+      {
+        profile: ProfileDto.new(@user),
+        timeline: PostDto.from_collection(posts)
+      }
+    end
+  end
+end
